Move tests to its own file
